Move get_world_map_stats to service layer
